create register.html

+++ Django/myProject/discord/views.py
@@ -3,6 +3,9 @@ from django.contrib import messages
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
+from django.contrib.auth.forms import UserCreationForm
 
 from .models import Room, Topic
 from .forms import RoomForm
@@ -21,8 +24,9 @@ from .forms import RoomForm
 
 
 def loginPage(request):
-    if request.method == "POST" :
-        username = request.POST.get('username')
+    page = 'login'
+    if request.method == "POST":
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')    
         
         try:
@@ -36,14 +40,33 @@ def loginPage(request):
             login(request, user)
             return redirect('home')
         else:
-            messages.error(request, 'Username or Password does not exit') 
+            messages.error(request, 'Username or Password does not exist') 
             
-    context = {}
+    context = {'page':page}
     return render(request, 'base/login_register.html', context)
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+def registerPage(request):
+    form = UserCreationForm()
+    
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        
+        else:
+            messages.error(request, 'Error occurred during registration')
+            
+
+    context = {'form':form}
+    return render(request, 'base/login_register.html', context)
 
 
 def home(request):
@@ -51,7 +74,8 @@ def home(request):
     rooms = Room.objects.filter(
         Q(topic__name__icontains=q) |
         Q(name__icontains=q) |
-        Q(description__icontains=q)
+        Q(description__icontains=q) |
+        Q(host__username__icontains=q) 
     )
     topics = Topic.objects.all()
     room_count = rooms.count()
@@ -63,6 +87,8 @@ def room(request, pk):
     context = {'room':room}
     return render(request, 'base/room.html', context)
 
+
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -74,10 +100,14 @@ def createRoom(request):
     context = {'form':form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    
+    if request.user != room.host:
+        return HttpResponse ("You are not allowed to Edit")
+    
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -88,9 +118,14 @@ def updateRoom(request, pk):
     context = {'form':form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def deleteRoom (request,pk):
     room = Room.objects.get(id=pk)
+    
+    if request.user != room.host:
+        return HttpResponse ("You are not allowed to Delete")
+    
+    
     if request.method == 'POST':
         room.delete()
         return redirect ('home')
